requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic
SQLAlchemy
python-multipart
